backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.26",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.30",
    "langchain-tavily>=0.2.13",
    "langgraph>=0.5.3",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "motor>=3.7.1",
    "openai>=1.97.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "pymongo==4.12.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.35.0",
]
